@@ -2,16 +2,16 @@ name: flutter_local_notifications
 description: A cross platform plugin for displaying and scheduling local
   notifications for Flutter applications with the ability to customise for each
   platform.
-version: 9.9.1
-repository: https://github.com/MaikuB/flutter_local_notifications/tree/master/flutter_local_notifications
+version: 10.0.0
+homepage: https://github.com/MaikuB/flutter_local_notifications/tree/master/flutter_local_notifications
 issue_tracker: https://github.com/MaikuB/flutter_local_notifications/issues
 
 dependencies:
   clock: ^1.1.0
   flutter:
     sdk: flutter
-  flutter_local_notifications_linux: ^0.5.0+1
-  flutter_local_notifications_platform_interface: ^5.0.0
+  flutter_local_notifications_linux: ^1.0.0
+  flutter_local_notifications_platform_interface: ^6.0.0
   timezone: ^0.8.0
 
 dev_dependencies:
@@ -36,5 +36,5 @@ flutter:
         default_package: flutter_local_notifications_linux
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.2.0"
+  sdk: ">=2.15.0 <3.0.0"
+  flutter: ">=2.8.0"
