@@ -1,12 +1,12 @@
 name: flutter_local_notifications_linux
 description: Linux implementation of the flutter_local_notifications plugin
-version: 0.5.1
+version: 1.0.0
 homepage: https://github.com/MaikuB/flutter_local_notifications/tree/master/flutter_local_notifications
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_local_notifications_platform_interface: ^5.0.0
+  flutter_local_notifications_platform_interface: ^6.0.0
   dbus: ^0.7.1
   path: ^1.8.0
   xdg_directories: ^0.2.0+1
@@ -17,5 +17,5 @@ dev_dependencies:
   mocktail: ^0.3.0
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.2.0"
+  sdk: ">=2.15.0 <3.0.0"
+  flutter: ">=2.8.0"
