@@ -1,18 +1,22 @@
 export 'package:flutter_local_notifications_linux/flutter_local_notifications_linux.dart';
 export 'package:flutter_local_notifications_platform_interface/flutter_local_notifications_platform_interface.dart'
     show
-        SelectNotificationCallback,
+        DidReceiveBackgroundNotificationResponseCallback,
+        DidReceiveNotificationResponseCallback,
         PendingNotificationRequest,
+        ActiveNotification,
         RepeatInterval,
-        NotificationAppLaunchDetails;
+        NotificationAppLaunchDetails,
+        NotificationResponse,
+        NotificationResponseType;
 
 export 'src/flutter_local_notifications_plugin.dart';
 export 'src/initialization_settings.dart';
 export 'src/notification_details.dart';
 export 'src/platform_flutter_local_notifications.dart'
     hide MethodChannelFlutterLocalNotificationsPlugin;
-export 'src/platform_specifics/android/active_notification.dart';
 export 'src/platform_specifics/android/bitmap.dart';
+export 'src/platform_specifics/android/categories.dart';
 export 'src/platform_specifics/android/enums.dart'
     hide AndroidBitmapSource, AndroidIconSource, AndroidNotificationSoundSource;
 export 'src/platform_specifics/android/icon.dart' hide AndroidIcon;
@@ -30,12 +34,14 @@ export 'src/platform_specifics/android/styles/inbox_style_information.dart';
 export 'src/platform_specifics/android/styles/media_style_information.dart';
 export 'src/platform_specifics/android/styles/messaging_style_information.dart';
 export 'src/platform_specifics/android/styles/style_information.dart';
+export 'src/platform_specifics/darwin/initialization_settings.dart';
+export 'src/platform_specifics/darwin/interruption_level.dart';
+export 'src/platform_specifics/darwin/notification_action.dart';
+export 'src/platform_specifics/darwin/notification_action_option.dart';
+export 'src/platform_specifics/darwin/notification_attachment.dart';
+export 'src/platform_specifics/darwin/notification_category.dart';
+export 'src/platform_specifics/darwin/notification_category_option.dart';
+export 'src/platform_specifics/darwin/notification_details.dart';
 export 'src/platform_specifics/ios/enums.dart';
-export 'src/platform_specifics/ios/initialization_settings.dart';
-export 'src/platform_specifics/ios/notification_attachment.dart';
-export 'src/platform_specifics/ios/notification_details.dart';
-export 'src/platform_specifics/macos/initialization_settings.dart';
-export 'src/platform_specifics/macos/notification_attachment.dart';
-export 'src/platform_specifics/macos/notification_details.dart';
 export 'src/typedefs.dart';
 export 'src/types.dart';
