@@ -1,8 +1,7 @@
 import 'package:flutter_local_notifications_linux/flutter_local_notifications_linux.dart';
 
 import 'platform_specifics/android/initialization_settings.dart';
-import 'platform_specifics/ios/initialization_settings.dart';
-import 'platform_specifics/macos/initialization_settings.dart';
+import 'platform_specifics/darwin/initialization_settings.dart';
 
 /// Settings for initializing the plugin for each platform.
 class InitializationSettings {
@@ -24,13 +23,13 @@ class InitializationSettings {
   ///
   /// It is nullable, because we don't want to force users to specify settings
   /// for platforms that they don't target.
-  final IOSInitializationSettings? iOS;
+  final DarwinInitializationSettings? iOS;
 
   /// Settings for macOS.
   ///
   /// It is nullable, because we don't want to force users to specify settings
   /// for platforms that they don't target.
-  final MacOSInitializationSettings? macOS;
+  final DarwinInitializationSettings? macOS;
 
   /// Settings for Linux.
   ///
