@@ -1,8 +1,7 @@
 import 'package:flutter_local_notifications_linux/flutter_local_notifications_linux.dart';
 
 import 'platform_specifics/android/notification_details.dart';
-import 'platform_specifics/ios/notification_details.dart';
-import 'platform_specifics/macos/notification_details.dart';
+import 'platform_specifics/darwin/notification_details.dart';
 
 /// Contains notification details specific to each platform.
 class NotificationDetails {
@@ -18,10 +17,10 @@ class NotificationDetails {
   final AndroidNotificationDetails? android;
 
   /// Notification details for iOS.
-  final IOSNotificationDetails? iOS;
+  final DarwinNotificationDetails? iOS;
 
   /// Notification details for macOS.
-  final MacOSNotificationDetails? macOS;
+  final DarwinNotificationDetails? macOS;
 
   /// Notification details for Linux.
   final LinuxNotificationDetails? linux;
