@@ -1,7 +1,8 @@
-/// Represents an attachment for an macOS notification.
-class MacOSNotificationAttachment {
-  /// Constructs an instance of [MacOSNotificationAttachment].
-  const MacOSNotificationAttachment(
+/// Represents an attachment for a notification on Darwin-based operation
+/// systems such as iOS and macOS
+class DarwinNotificationAttachment {
+  /// Constructs an instance of [DarwinNotificationAttachment].
+  const DarwinNotificationAttachment(
     this.filePath, {
     this.identifier,
   });
@@ -14,6 +15,7 @@ class MacOSNotificationAttachment {
 
   /// The unique identifier for the attachment.
   ///
-  /// When left empty, the macOS APIs will generate a unique identifier
+  /// When left empty, the platform's native APIs will generate a unique
+  /// identifier
   final String? identifier;
 }
