@@ -15,13 +15,16 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
         static let requestAlertPermission = "requestAlertPermission"
         static let requestSoundPermission = "requestSoundPermission"
         static let requestBadgePermission = "requestBadgePermission"
+        static let requestCriticalPermission = "requestCriticalPermission"
         static let alert = "alert"
         static let sound = "sound"
         static let badge = "badge"
+        static let critical = "critical"
         static let notificationLaunchedApp = "notificationLaunchedApp"
         static let id = "id"
         static let title = "title"
         static let subtitle = "subtitle"
+        static let categoryIdentifier = "categoryIdentifier"
         static let body = "body"
         static let scheduledDateTime = "scheduledDateTime"
         static let timeZoneName = "timeZoneName"
@@ -33,6 +36,17 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
         static let identifier = "identifier"
         static let filePath = "filePath"
         static let threadIdentifier = "threadIdentifier"
+        static let interruptionLevel = "interruptionLevel"
+        static let actionId = "actionId"
+        static let notificationResponseType = "notificationResponseType"
+    }
+
+    struct ErrorMessages {
+        static let getActiveNotificationsErrorMessage = "macOS version must be 10.14 or newer to use getActiveNotifications"
+    }
+
+    struct ErrorCodes {
+        static let unsupportedOSVersion = "unsupported_os_version"
     }
 
     struct DateFormatStrings {
@@ -63,7 +77,7 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
     var defaultPresentAlert = false
     var defaultPresentSound = false
     var defaultPresentBadge = false
-    var launchPayload: String?
+    var launchNotificationResponseDict: [String: Any?]?
     var launchingAppFromNotification = false
 
     init(fromChannel channel: FlutterMethodChannel) {
@@ -109,18 +123,36 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
         if !isAFlutterLocalNotification(userInfo: response.notification.request.content.userInfo) {
             return
         }
-        let payload = response.notification.request.content.userInfo[MethodCallArguments.payload] as? String
-        if initialized {
-            handleSelectNotification(payload: payload)
+        if response.actionIdentifier == UNNotificationDefaultActionIdentifier {
+            let payload = response.notification.request.content.userInfo[MethodCallArguments.payload] as? String
+            if initialized {
+                handleSelectNotification(notificationId: Int(response.notification.request.identifier)!, payload: payload)
+            } else {
+                launchNotificationResponseDict = extractNotificationResponseDict(response: response)
+                launchingAppFromNotification = true
+            }
+
+            completionHandler()
+        } else if response.actionIdentifier == UNNotificationDismissActionIdentifier {
+            completionHandler()
         } else {
-            launchPayload = payload
-            launchingAppFromNotification = true
+            if initialized {
+                // No isolate can be used for macOS until https://github.com/flutter/flutter/issues/65222 is resolved.
+                //
+                // Therefore, we call the regular method channel and let the macos plugin handle it appropriately.
+                handleSelectNotificationAction(arguments: extractNotificationResponseDict(response: response))
+            } else {
+                launchNotificationResponseDict = extractNotificationResponseDict(response: response)
+                launchingAppFromNotification = true
+            }
+
+            completionHandler()
         }
     }
 
     public func userNotificationCenter(_ center: NSUserNotificationCenter, didActivate notification: NSUserNotification) {
         if notification.activationType == .contentsClicked && notification.userInfo != nil && isAFlutterLocalNotification(userInfo: notification.userInfo!) {
-            handleSelectNotification(payload: notification.userInfo![MethodCallArguments.payload] as? String)
+            handleSelectNotification(notificationId: Int(notification.identifier!)!, payload: notification.userInfo![MethodCallArguments.payload] as? String)
         }
     }
 
@@ -142,6 +174,8 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
             cancelAll(result)
         case "pendingNotificationRequests":
             pendingNotificationRequests(result)
+        case "getActiveNotifications":
+            getActiveNotifications(result)
         case "show":
             show(call, result)
         case "zonedSchedule":
@@ -162,7 +196,18 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
             let requestedAlertPermission = arguments[MethodCallArguments.requestAlertPermission] as! Bool
             let requestedSoundPermission = arguments[MethodCallArguments.requestSoundPermission] as! Bool
             let requestedBadgePermission = arguments[MethodCallArguments.requestBadgePermission] as! Bool
-            requestPermissionsImpl(soundPermission: requestedSoundPermission, alertPermission: requestedAlertPermission, badgePermission: requestedBadgePermission, result: result)
+            let requestedCriticalPermission = arguments[MethodCallArguments.requestCriticalPermission] as! Bool
+
+            configureNotificationCategories(arguments) {
+                self.requestPermissionsImpl(
+                    soundPermission: requestedSoundPermission,
+                    alertPermission: requestedAlertPermission,
+                    badgePermission: requestedBadgePermission,
+                    criticalPermission: requestedCriticalPermission,
+                    result: result
+                )
+            }
+
             initialized = true
         } else {
             result(true)
@@ -170,13 +215,76 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
         }
     }
 
+    func configureNotificationCategories(_ arguments: [String: AnyObject],
+                                         withCompletionHandler completionHandler: @escaping () -> Void) {
+        if #available(OSX 10.14, *) {
+            if let categories = arguments["notificationCategories"] as? [[String: AnyObject]] {
+                var notificationCategories = Set<UNNotificationCategory>()
+
+                for category in categories {
+                    var newActions = [UNNotificationAction]()
+
+                    if let actions = category["actions"] as? [[String: AnyObject]] {
+                        for action in actions {
+                            let type = action["type"] as! String
+                            let identifier = action["identifier"] as! String
+                            let title = action["title"] as! String
+                            let options = action["options"] as! [Any]
+                            if type == "plain" {
+                                newActions.append(UNNotificationAction(
+                                    identifier: identifier,
+                                    title: title,
+                                    options: Converters.parseNotificationActionOptions(options)
+                                ))
+                            } else if type == "text" {
+                                let buttonTitle = action["buttonTitle"] as! String
+                                let placeholder = action["placeholder"] as! String
+                                newActions.append(UNTextInputNotificationAction(
+                                    identifier: identifier,
+                                    title: title,
+                                    options: Converters.parseNotificationActionOptions(options),
+                                    textInputButtonTitle: buttonTitle,
+                                    textInputPlaceholder: placeholder
+                                ))
+                            }
+                        }
+                    }
+
+                    let notificationCategory = UNNotificationCategory(
+                        identifier: category["identifier"] as! String,
+                        actions: newActions,
+                        intentIdentifiers: [],
+                        hiddenPreviewsBodyPlaceholder: nil,
+                        categorySummaryFormat: nil,
+                        options: Converters.parseNotificationCategoryOptions(category["options"] as! [NSNumber])
+                    )
+
+                    notificationCategories.insert(notificationCategory)
+                }
+
+                if !notificationCategories.isEmpty {
+                    let center = UNUserNotificationCenter.current()
+                    center.setNotificationCategories(notificationCategories)
+                }
+
+                completionHandler()
+
+            } else {
+                completionHandler()
+            }
+        } else {
+            completionHandler()
+        }
+    }
+
     func requestPermissions(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
         if #available(OSX 10.14, *) {
             let arguments = call.arguments as! [String: AnyObject]
-            let requestedAlertPermission = arguments[MethodCallArguments.alert] as? Bool
-            let requestedSoundPermission = arguments[MethodCallArguments.sound] as? Bool
-            let requestedBadgePermission = arguments[MethodCallArguments.badge] as? Bool
-            requestPermissionsImpl(soundPermission: requestedSoundPermission ?? false, alertPermission: requestedAlertPermission ?? false, badgePermission: requestedBadgePermission ?? false, result: result)
+            let requestedAlertPermission = arguments[MethodCallArguments.alert] as! Bool
+            let requestedSoundPermission = arguments[MethodCallArguments.sound] as! Bool
+            let requestedBadgePermission = arguments[MethodCallArguments.badge] as! Bool
+            let requestedCriticalPermission = arguments[MethodCallArguments.critical] as! Bool
+            requestPermissionsImpl(soundPermission: requestedSoundPermission, alertPermission: requestedAlertPermission, badgePermission: requestedBadgePermission, criticalPermission: requestedCriticalPermission, result: result)
         } else {
             result(nil)
         }
@@ -184,7 +292,7 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
 
     func getNotificationAppLaunchDetails(_ result: @escaping FlutterResult) {
         if #available(OSX 10.14, *) {
-            let appLaunchDetails: [String: Any?] = [MethodCallArguments.notificationLaunchedApp: launchingAppFromNotification, MethodCallArguments.payload: launchPayload]
+            let appLaunchDetails: [String: Any?] = [MethodCallArguments.notificationLaunchedApp: launchingAppFromNotification, "notificationResponse": launchNotificationResponseDict]
             result(appLaunchDetails)
         } else {
             result(nil)
@@ -246,6 +354,20 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
                 requestDictionaries.append([MethodCallArguments.id: Int(scheduledNotification.identifier!) as Any, MethodCallArguments.title: scheduledNotification.title, MethodCallArguments.body: scheduledNotification.informativeText, MethodCallArguments.payload: scheduledNotification.userInfo![MethodCallArguments.payload]])
             }
             result(requestDictionaries)
+        }
+    }
+
+    func getActiveNotifications(_ result: @escaping FlutterResult) {
+        if #available(OSX 10.14, *) {
+            UNUserNotificationCenter.current().getDeliveredNotifications { (requests) in
+                var requestDictionaries: [[String: Any?]] = []
+                for request in requests {
+                    requestDictionaries.append([MethodCallArguments.id: Int(request.request.identifier) as Any, MethodCallArguments.title: request.request.content.title, MethodCallArguments.body: request.request.content.body, MethodCallArguments.payload: request.request.content.userInfo[MethodCallArguments.payload]])
+                }
+                result(requestDictionaries)
+            }
+        } else {
+            result(FlutterError.init(code: ErrorCodes.unsupportedOSVersion, message: ErrorMessages.getActiveNotificationsErrorMessage, details: nil))
         }
     }
 
@@ -382,6 +504,14 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
             if let threadIdentifier = platformSpecifics[MethodCallArguments.threadIdentifier] as? String {
                 content.threadIdentifier = threadIdentifier
             }
+            if let categoryIdentifier = platformSpecifics[MethodCallArguments.categoryIdentifier] as? String {
+                content.categoryIdentifier = categoryIdentifier
+            }
+            if #available(macOS 12.0, *) {
+              if let rawInterruptionLevel = platformSpecifics[MethodCallArguments.interruptionLevel] as? UInt {
+                content.interruptionLevel = UNNotificationInterruptionLevel.init(rawValue: rawInterruptionLevel)!
+              }
+            }
             if let attachments = platformSpecifics[MethodCallArguments.attachments] as? [[String: AnyObject]] {
                 content.attachments = []
                 for attachment in attachments {
@@ -454,8 +584,8 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
     }
 
     @available(OSX 10.14, *)
-    func requestPermissionsImpl(soundPermission: Bool, alertPermission: Bool, badgePermission: Bool, result: @escaping FlutterResult) {
-        if !soundPermission && !alertPermission && !badgePermission {
+    func requestPermissionsImpl(soundPermission: Bool, alertPermission: Bool, badgePermission: Bool, criticalPermission: Bool, result: @escaping FlutterResult) {
+        if !soundPermission && !alertPermission && !badgePermission && !criticalPermission {
             result(false)
             return
         }
@@ -468,6 +598,9 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
         }
         if badgePermission {
             options.insert(.badge)
+        }
+        if criticalPermission {
+            options.insert(.criticalAlert)
         }
         UNUserNotificationCenter.current().requestAuthorization(options: options) { (granted, _) in
             result(granted)
@@ -508,8 +641,31 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
         return String(arguments[MethodCallArguments.id] as! Int)
     }
 
-    func handleSelectNotification(payload: String?) {
-        channel.invokeMethod("selectNotification", arguments: payload)
+    func handleSelectNotification(notificationId: Int, payload: String?) {
+        var arguments: [String: Any?] = [:]
+        arguments["notificationId"] = notificationId
+        arguments["payload"] = payload
+        arguments["notificationResponseType"] = 0
+        channel.invokeMethod("didReceiveNotificationResponse", arguments: arguments)
+    }
+
+    func handleSelectNotificationAction(arguments: [String: Any?]) {
+        channel.invokeMethod("didReceiveNotificationResponse", arguments: arguments)
+    }
+
+    @available(macOS 10.14, *)
+    func extractNotificationResponseDict(response: UNNotificationResponse) -> [String: Any?] {
+        var notificationResponseDict: [String: Any?] = [:]
+        notificationResponseDict["notificationId"] = Int(response.notification.request.identifier)!
+        if response.actionIdentifier == UNNotificationDefaultActionIdentifier {
+            notificationResponseDict[MethodCallArguments.notificationResponseType] = 0
+        } else if response.actionIdentifier != UNNotificationDismissActionIdentifier {
+            notificationResponseDict[MethodCallArguments.actionId] = response.actionIdentifier
+            notificationResponseDict[MethodCallArguments.notificationResponseType] = 1
+        }
+        notificationResponseDict["input"] = (response as? UNTextInputNotificationResponse)?.userText
+        notificationResponseDict[MethodCallArguments.payload] = response.notification.request.content.userInfo[MethodCallArguments.payload]
+        return notificationResponseDict
     }
 
     func isAFlutterLocalNotification(userInfo: [AnyHashable: Any]) -> Bool {
