@@ -12,8 +12,8 @@ void main() {
   FlutterLocalNotificationsPlatform.instance = mock;
 
   test('Creates IOSNotificationAttachment when file path is specified', () {
-    expect(
-        const IOSNotificationAttachment(''), isA<IOSNotificationAttachment>());
+    expect(const DarwinNotificationAttachment(''),
+        isA<DarwinNotificationAttachment>());
   });
 }
 
