@@ -9,94 +9,93 @@ class LinuxNotificationCategory {
 
   /// A generic device-related notification
   /// that doesn't fit into any other category.
-  factory LinuxNotificationCategory.device() =>
-      const LinuxNotificationCategory('device');
+  static const LinuxNotificationCategory device =
+      LinuxNotificationCategory('device');
 
   /// A device, such as a USB device, was added to the system.
-  factory LinuxNotificationCategory.deviceAdded() =>
-      const LinuxNotificationCategory('device.added');
+  static const LinuxNotificationCategory deviceAdded =
+      LinuxNotificationCategory('device.added');
 
   /// A device had some kind of error.
-  factory LinuxNotificationCategory.deviceError() =>
-      const LinuxNotificationCategory('device.error');
+  static const LinuxNotificationCategory deviceError =
+      LinuxNotificationCategory('device.error');
 
   /// A device, such as a USB device, was removed from the system.
-  factory LinuxNotificationCategory.deviceRemoved() =>
-      const LinuxNotificationCategory('device.removed');
+  static const LinuxNotificationCategory deviceRemoved =
+      LinuxNotificationCategory('device.removed');
 
   /// A generic e-mail-related notification
   /// that doesn't fit into any other category.
-  factory LinuxNotificationCategory.email() =>
-      const LinuxNotificationCategory('email');
+  static const LinuxNotificationCategory email =
+      LinuxNotificationCategory('email');
 
   /// A new e-mail notification.
-  factory LinuxNotificationCategory.emailArrived() =>
-      const LinuxNotificationCategory('email.arrived');
+  static const LinuxNotificationCategory emailArrived =
+      LinuxNotificationCategory('email.arrived');
 
   /// A notification stating that an e-mail has bounced.
-  factory LinuxNotificationCategory.emailBounced() =>
-      const LinuxNotificationCategory('email.bounced');
+  static const LinuxNotificationCategory emailBounced =
+      LinuxNotificationCategory('email.bounced');
 
   /// A generic instant message-related notification
   /// that doesn't fit into any other
-  factory LinuxNotificationCategory.im() =>
-      const LinuxNotificationCategory('im');
+  static const LinuxNotificationCategory im = LinuxNotificationCategory('im');
 
   /// An instant message error notification.
-  factory LinuxNotificationCategory.imError() =>
-      const LinuxNotificationCategory('im.error');
+  static const LinuxNotificationCategory imError =
+      LinuxNotificationCategory('im.error');
 
   /// A received instant message notification.
-  factory LinuxNotificationCategory.imReceived() =>
-      const LinuxNotificationCategory('im.received');
+  static const LinuxNotificationCategory imReceived =
+      LinuxNotificationCategory('im.received');
 
   /// A generic network notification that
   /// doesn't fit into any other category.
-  factory LinuxNotificationCategory.network() =>
-      const LinuxNotificationCategory('network');
+  static const LinuxNotificationCategory network =
+      LinuxNotificationCategory('network');
 
   /// A network connection notification,
   /// such as successful sign-on to a network service.
   /// This should not be confused with
-  /// [LinuxNotificationCategory.deviceAdded] for new network devices.
-  factory LinuxNotificationCategory.networkConnected() =>
-      const LinuxNotificationCategory('network.connected');
+  /// [deviceAdded] for new network devices.
+  static const LinuxNotificationCategory networkConnected =
+      LinuxNotificationCategory('network.connected');
 
   /// A network disconnected notification.
-  /// This should not be confused with [LinuxNotificationCategory.deviceRemoved]
+  /// This should not be confused with [deviceRemoved]
   /// for disconnected network devices.
-  factory LinuxNotificationCategory.networkDisconnected() =>
-      const LinuxNotificationCategory('network.disconnected');
+  static const LinuxNotificationCategory networkDisconnected =
+      LinuxNotificationCategory('network.disconnected');
 
   /// A network-related or connection-related error.
-  factory LinuxNotificationCategory.networkError() =>
-      const LinuxNotificationCategory('network.error');
+  static const LinuxNotificationCategory networkError =
+      LinuxNotificationCategory('network.error');
 
   /// A generic presence change notification
   /// that doesn't fit into any other category, such as going away or idle.
-  factory LinuxNotificationCategory.presence() =>
-      const LinuxNotificationCategory('presence');
+  static const LinuxNotificationCategory presence =
+      LinuxNotificationCategory('presence');
 
   /// An offline presence change notification.
-  factory LinuxNotificationCategory.presenceOffile() =>
-      const LinuxNotificationCategory('presence.offline');
+  static const LinuxNotificationCategory presenceOffile =
+      LinuxNotificationCategory('presence.offline');
 
   /// An online presence change notification.
-  factory LinuxNotificationCategory.presenceOnline() =>
-      const LinuxNotificationCategory('presence.online');
+  static const LinuxNotificationCategory presenceOnline =
+      LinuxNotificationCategory('presence.online');
 
   /// A generic file transfer or download notification
   /// that doesn't fit into any other category.
-  factory LinuxNotificationCategory.transfer() =>
-      const LinuxNotificationCategory('transfer');
+  static const LinuxNotificationCategory transfer =
+      LinuxNotificationCategory('transfer');
 
   /// A file transfer or download complete notification.
-  factory LinuxNotificationCategory.transferComplete() =>
-      const LinuxNotificationCategory('transfer.complete');
+  static const LinuxNotificationCategory transferComplete =
+      LinuxNotificationCategory('transfer.complete');
 
   /// A file transfer or download error.
-  factory LinuxNotificationCategory.transferError() =>
-      const LinuxNotificationCategory('transfer.error');
+  static const LinuxNotificationCategory transferError =
+      LinuxNotificationCategory('transfer.error');
 
   /// Name of category.
   final String name;
