@@ -3,5 +3,5 @@ export 'mock_dbus_wrapper.dart';
 export 'mock_file.dart';
 export 'mock_file_system.dart';
 export 'mock_platform_info.dart';
-export 'mock_select_notification_callback.dart';
 export 'mock_storage.dart';
+export 'mock_typedef.dart';
